Feature/fin 034

+++ src/server/shared/infrastructure/crud.api.repository.ts
@@ -1,8 +1,8 @@
-import { EntityTarget, FindOptionsWhere, ObjectLiteral } from 'typeorm';
-import { DefaultColumnKeys, DefaultTableColumns } from '@common/models/default-table-columns.model';
-import { ICrudService } from '@common/models/crud-service.model';
+import { type EntityTarget, type FindOptionsWhere, type ObjectLiteral } from 'typeorm';
+import { type DefaultColumnKeys, type DefaultTableColumns } from '@common/models/default-table-columns.model';
+import { type ICrudService } from '@common/models/crud-service.model';
 import { OrmRepository } from './orm.repository';
-import { DeepPartial } from '@common/models/deep-partial.model';
+import { type DeepPartial } from '@common/models/deep-partial.model';
 
 export abstract class CrudApiRepository<
   T extends ObjectLiteral & DefaultTableColumns,
@@ -45,7 +45,13 @@ export abstract class CrudApiRepository<
     return this.repository.find({ skip, take, where: where as FindOptionsWhere<T> });
   }
 
-  async deleteItem(id: number): Promise<true> {
+  async deleteItem(id: number, softDeleted?: boolean): Promise<true> {
+    if (softDeleted) {
+      await this.repository.update({ id } as FindOptionsWhere<T>, { softDeleted: 1 } as T);
+
+      return true;
+    }
+
     await this.repository.delete({ id } as FindOptionsWhere<T>);
 
     return true;
